修复 NewDamageModel 中 MathF.Max 类型不匹配

+++ 工程/New_ZZZF/Systems/NewDamageModel.cs
@@ -537,7 +537,7 @@ namespace New_ZZZF
 
             num = MathF.Max(
                 num,
-                weapon.GetModifiedThrustDamageForCurrentUsage());
+                (float)weapon.GetModifiedThrustDamageForCurrentUsage());
 
             if (item.IsCraftedByPlayer)
             {
@@ -679,7 +679,7 @@ namespace New_ZZZF
 
             num = MathF.Max(
                 num,
-                weapon.GetModifiedThrustDamageForCurrentUsage());
+                (float)weapon.GetModifiedThrustDamageForCurrentUsage());
 
             string? affixInstId = AffixMissionBehavior.GetAgentWeaponInstanceId(
                 attackInformation.AttackerAgent,
